// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A scalar that can represent any JSON Object value. */
  JSONObject: any;
  RempReceiptJson: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export type BlockSignatures = {
  __typename?: 'BlockSignatures';
  block?: Maybe<Block>;
  /**
   *
   * Complexity 0
   */
  catchain_seqno?: Maybe<Scalars['Float']>;
  /**
   * Signed block's gen_utime
   *
   * Complexity 0
   */
  gen_utime?: Maybe<Scalars['Float']>;
  /**
   *
   * Complexity 0
   */
  gen_utime_string?: Maybe<Scalars['String']>;
  /**
   *
   * Complexity 0
   */
  id?: Maybe<Scalars['String']>;
  /**
   * Signed block's merkle proof
   *
   * Complexity 0
   */
  proof?: Maybe<Scalars['String']>;
  /**
   * Signed block's seq_no
   *
   * Complexity 0
   */
  seq_no?: Maybe<Scalars['Float']>;
  /**
   * Signed block's shard
   *
   * Complexity 0
   */
  shard?: Maybe<Scalars['String']>;
  /**
   *
   * Complexity 0
   */
  sig_weight?: Maybe<Scalars['String']>;
  /** Array of signatures from block's validators */
  signatures?: Maybe<Array<Maybe<BlockSignaturesSignatures>>>;
  /**
   *
   * Complexity 0
   */
  validator_list_hash_short?: Maybe<Scalars['Float']>;
  /**
   * Signed block's workchain_id
   *
   * Complexity 0
   */
  workchain_id?: Maybe<Scalars['Int']>;
};
// ...
export type BlockchainAccount = Node & {
  __typename?: 'BlockchainAccount';
  /**
   * Returns the current status of the account.
   *
   * - 0 – uninit
   * - 1 – active
   * - 2 – frozen
   * - 3 – nonExist
   *
   * Complexity 0
   */
  acc_type?: Maybe<Scalars['Int']>;
  acc_type_name?: Maybe<AccountStatusEnum>;
  /**
   *
   * Complexity 0
   */
  address?: Maybe<Scalars['String']>;
  /**
   *
   * Complexity 0
   */
  balance?: Maybe<Scalars['String']>;
  balance_other?: Maybe<Array<Maybe<OtherCurrency>>>;
  /**
   * Contains sum of all the bits used by the cells of the account. Used in storage fee calculation
   *
   * Complexity 0
   */
  bits?: Maybe<Scalars['String']>;
  /**
   * Bag of cells with the account struct encoded as base64.
   *
   * Complexity 5
   */
  boc?: Maybe<Scalars['String']>;
  /**
   * Contains number of the cells of the account. Used in storage fee calculation
   *
   * Complexity 0
   */
  cells?: Maybe<Scalars['String']>;
  /**
   * If present, contains smart-contract code encoded with in base64.
   *
   * Complexity 2
   */
  code?: Maybe<Scalars['String']>;
  /**
   * `code` field root hash.
   *
   * Complexity 0
   */
  code_hash?: Maybe<Scalars['String']>;
  /**
   * If present, contains smart-contract data encoded with in base64.
   *
   * Complexity 3
   */
  data?: Maybe<Scalars['String']>;
  /**
   * `data` field root hash.
   *
   * Complexity 0
   */
  data_hash?: Maybe<Scalars['String']>;
  /**
   * If present, accumulates the storage payments that could not be exacted from
   * the balance of the account, represented by a strictly positive amount of nano
   * tokens; it can be present only for uninitialized or frozen accounts that have
   * a balance of zero Grams (but may have non-zero balances in non gram
   * cryptocurrencies). When due_payment becomes larger than the value of a
   * configurable parameter of the blockchain, the ac- count is destroyed
   * altogether, and its balance, if any, is transferred to the zero account.
   *
   * Complexity 0
   */
  due_payment?: Maybe<Scalars['String']>;
  /**
   * BlockchainAccount.id is "account/"-prefixed Account.id.
   * For id without prefix see "address".
   *
   * Complexity 0
   */
  id: Scalars['ID'];
  /**
   * account 's initial code hash (when it was deployed)
   *
   * Complexity 0
   */
  init_code_hash?: Maybe<Scalars['String']>;
  /**
   * Contains either the unixtime of the most recent storage payment
   * collected (usually this is the unixtime of the most recent transaction),
   * or the unixtime when the account was created (again, by a transaction).
   *
   * Complexity 0
   */
  last_paid?: Maybe<Scalars['Float']>;
  /**
   *
   * Complexity 0
   */
  last_trans_lt?: Maybe<Scalars['String']>;
  /**
   * If present, contains library code used in smart-contract.
   *
   * Complexity 0
   */
  library?: Maybe<Scalars['String']>;
  /**
   * `library` field root hash.
   *
   * Complexity 0
   */
  library_hash?: Maybe<Scalars['String']>;
  /**
   *
   * Complexity 0
   */
  prev_code_hash?: Maybe<Scalars['String']>;
  /**
   * Merkle proof that account is a part of shard state it cut from as a bag of cells with Merkle proof struct encoded as base64.
   *
   * Complexity 0
   */
  proof?: Maybe<Scalars['String']>;
  /**
   * Contains the number of public cells of the account. Used in storage fee calculation.
   *
   * Complexity 0
   */
  public_cells?: Maybe<Scalars['String']>;
  /**
   * Is present and non-zero only in instances of large smart contracts.
   *
   * Complexity 0
   */
  split_depth?: Maybe<Scalars['Int']>;
  /**
   * Contains the representation hash of an instance of `StateInit` when an account is frozen.
   *
   * Complexity 0
   */
  state_hash?: Maybe<Scalars['String']>;
  /**
   * May be present only in the masterchain—and within the masterchain, only in some
   * fundamental smart contracts required for the whole system to function.
   *
   * Complexity 0
   */
  tick?: Maybe<Scalars['Boolean']>;
  /**
   * May be present only in the masterchain—and within the masterchain, only in some
   * fundamental smart contracts required for the whole system to function.
   *
   * Complexity 0
   */
  tock?: Maybe<Scalars['Boolean']>;
  tokenHolder?: Maybe<FungibleTokenHolder>;
  /**
   * Workchain id of the account address (id field).
   *
   * Complexity 0
   */
  workchain_id?: Maybe<Scalars['Int']>;
};
// ...
export type BlockchainQuery = {
  __typename?: 'BlockchainQuery';
  /** Account-related information */
  account?: Maybe<BlockchainAccountQuery>;
  block?: Maybe<BlockchainBlock>;
  block_by_seq_no?: Maybe<BlockchainBlock>;
  /** This node could be used for a cursor-based pagination of blocks. */
  blocks?: Maybe<BlockchainBlocksConnection>;
  /** This node could be used for a cursor-based pagination of key blocks. */
  key_blocks?: Maybe<BlockchainBlocksConnection>;
  /**
   * Returns masterchain seq_no range for the specified time range
   * to be used further in pagination functions.
   * If `time_start` and/or `time_end` is null, then the corresponding seq_no range border
   * is also null.
   */
  master_seq_no_range?: Maybe<BlockchainMasterSeqNoRange>;
  message?: Maybe<BlockchainMessage>;
  transaction?: Maybe<BlockchainTransaction>;
  /** This node could be used for a cursor-based pagination of transactions. */
  transactions?: Maybe<BlockchainTransactionsConnection>;
};
// ...
export type BlockchainTransactionEdge = {
  __typename?: 'BlockchainTransactionEdge';
  /**
   *
   * Complexity 0
   */
  cursor: Scalars['String'];
  node: BlockchainTransaction;
};

/**
 *
 * Complexity 0
 */
export type BlockchainTransactionsConnection = {
  __typename?: 'BlockchainTransactionsConnection';
  edges: Array<BlockchainTransactionEdge>;
  pageInfo: PageInfo;
};
// ...
export type Cost = {
  __typename?: 'Cost';
  /**
   *
   * Complexity 0
   */
  complexity?: Maybe<Scalars['Int']>;
  /**
   *
   * Complexity 0
   */
  explain?: Maybe<Array<Maybe<Scalars['String']>>>;
};

export enum CountByVersionPeriod {
  Day = 'DAY',
  Hour = 'HOUR'
}

/**
 * Counterparty
 *
 * Complexity 20
 */
export type Counterparty = {
  __typename?: 'Counterparty';
  /**
   *
   * Complexity 0
   */
  account?: Maybe<Scalars['String']>;
  /**
   *
   * Complexity 0
   */
  counterparty?: Maybe<Scalars['String']>;
  /**
   *
   * Complexity 0
   */
  cursor?: Maybe<Scalars['String']>;
  /**
   *
   * Complexity 0
   */
  last_message_at?: Maybe<Scalars['Float']>;
  /**
   *
   * Complexity 0
   */
  last_message_id?: Maybe<Scalars['String']>;
  /**
   *
   * Complexity 0
   */
  last_message_is_reverse?: Maybe<Scalars['Boolean']>;
  /**
   *
   * Complexity 0
   */
  last_message_value?: Maybe<Scalars['String']>;
};
// ...
export type FungibleTokenQuery = {
  __typename?: 'FungibleTokenQuery';
  /** Returns Holder */
  holder: FungibleTokenHolder;
  /** Returns holders by addresses */
  holdersByAddresses: Array<FungibleTokenHolder>;
  /** Statistics over all fungible token standarts */
  statistics: Array<FungibleTokenStandardStatistics>;
  /** Returns Token */
  token?: Maybe<FungibleToken>;
  /** Returns Tokens */
  tokens: FungibleTokenConnection;
  /** Returns transfers by message ids */
  transfersByMessageIds: Array<FungibleTokenTransfer>;
  /**
   * Stat API version
   *
   * Complexity 0
   */
  version: Scalars['String'];
  /** Returns wallet */
  wallet?: Maybe<FungibleTokenWallet>;
};
// ...
export type Node = {
  id: Scalars['ID'];
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /**
   *
   * Complexity 0
   */
  endCursor?: Maybe<Scalars['String']>;
  /**
   *
   * Complexity 0
   */
  hasNextPage: Scalars['Boolean'];
  /**
   *
   * Complexity 0
   */
  hasPreviousPage: Scalars['Boolean'];
  /**
   *
   * Complexity 0
   */
  startCursor?: Maybe<Scalars['String']>;
};

/**
 *
 * Complexity 0
 */
export type Price = {
  __typename?: 'Price';
  marketCap: EverMarketCap;
  price: EverPrice;
  /**
   *
   * Complexity 0
   */
  updatedAt: Scalars['Float'];
  /**
   *
   * Complexity 0
   */
  updatedISO: Scalars['String'];
  volume: EverVolume;
};

export type Query = {
  __typename?: 'Query';
  accounts?: Maybe<Array<Maybe<Account>>>;
  /**
   * Compute aggregated values for specified fields
   *
   * Complexity 100
   */
  aggregateAccounts?: Maybe<Array<Maybe<Scalars['String']>>>;
  /**
   * Compute aggregated values for specified fields
   *
   * Complexity 100
   */
  aggregateBlockSignatures?: Maybe<Array<Maybe<Scalars['String']>>>;
  /**
   * Compute aggregated values for specified fields
   *
   * Complexity 100
   */
  aggregateBlocks?: Maybe<Array<Maybe<Scalars['String']>>>;
  /**
   * Compute aggregated values for specified fields
   *
   * Complexity 100
   */
  aggregateMessages?: Maybe<Array<Maybe<Scalars['String']>>>;
  /**
   * Compute aggregated values for specified fields
   *
   * Complexity 100
   */
  aggregateTransactions?: Maybe<Array<Maybe<Scalars['String']>>>;
  /** Blockchain-related information (blocks, transactions, etc.) */
  blockchain?: Maybe<BlockchainQuery>;
  blocks?: Maybe<Array<Maybe<Block>>>;
  blocks_signatures?: Maybe<Array<Maybe<BlockSignatures>>>;
  cost?: Maybe<Cost>;
  /** Returns account counterparties */
  counterparties?: Maybe<Array<Maybe<Counterparty>>>;
  /** Flex API */
  flex: FlexQuery;
  /** FT API */
  ft: FungibleTokenQuery;
  info?: Maybe<Info>;
  /**
   * Version of the service that exposes recentExtInMessageStatuses query and
   * recentExtInMessageStatuses subscription
   *
   * Complexity 0
   */
  messageMonitorVersion: Scalars['String'];
  messages?: Maybe<Array<Maybe<Message>>>;
  price: Price;
  /**
   * Returns current processing status for each message.
   * Returns the result transaction if the message was successfully processed, in case of TIMEOUT
   * tries to return the result of transaction emulation.
   */
  recentExtInMessageStatuses: Array<MessageMonitoringResult>;
  /** Statistics API */
  statistics: StatisticsQuery;
  transactions?: Maybe<Array<Maybe<Transaction>>>;
  zerostates?: Maybe<Array<Maybe<Zerostate>>>;
};
// ...
export type GetLatestTransactionsQuery = { __typename?: 'Query', blockchain?: { __typename?: 'BlockchainQuery', transactions?: { __typename?: 'BlockchainTransactionsConnection', edges: Array<{ __typename?: 'BlockchainTransactionEdge', node: { __typename?: 'BlockchainTransaction', block_id?: string | null, workchain_id?: number | null, id: string, account?: { __typename?: 'BlockchainAccount', address?: string | null } | null } }>, pageInfo: { __typename?: 'PageInfo', endCursor?: string | null, startCursor?: string | null, hasPreviousPage: boolean, hasNextPage: boolean } } | null } | null };